task_backend/requirements.txt
# =====================================================
# Web Framework
# =====================================================
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0

# =====================================================
# Validation & Settings
# =====================================================
pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0,<2.0

# =====================================================
# Databases & Cache
# =====================================================
motor>=3.3,<4.0
redis>=5.0,<6.0

# =====================================================
# Authentication & Security
# =====================================================
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
python-multipart>=0.0.9,<1.0

# =====================================================
# Performance
# =====================================================
orjson>=3.9,<4.0

# =====================================================
# Testing
# =====================================================
pytest>=8.0,<10.0
pytest-asyncio>=0.23,<2.0
httpx>=0.27,<1.0

# =====================================================
# Google Cloud / Vertex AI
# =====================================================
google-cloud-aiplatform>=1.38.0,<2.0
vertexai>=1.43.0,<2.0
google-auth>=2.23.0,<3.0

# =====================================================
# AI / Math
# =====================================================
numpy>=1.26,<2.0

# =====================================================
# Data Processing
# =====================================================
pandas>=2.1,<3.0
openpyxl>=3.1,<4.0

# =====================================================
# PDF Processing
# =====================================================
pypdf>=4.0,<7.0

# =====================================================
# Async / Networking
# =====================================================
requests>=2.31.0,<3.0
aiofiles>=23.2.1,<24.0

# =====================================================
# ClickHouse
# =====================================================
clickhouse-driver==0.2.6
clickhouse-connect==0.7.0

# =====================================================
# MySQL
# =====================================================
pymysql==1.1.0
sshtunnel==0.4.0